+++ src/main/java/ee/carlrobert/codegpt/settings/configuration/ConfigurationState.java
@@ -20,7 +20,6 @@ public class ConfigurationState {
   private boolean methodNameGenerationEnabled = true;
   private boolean captureCompileErrors = true;
   private boolean autoFormattingEnabled = true;
-  private boolean codeCompletionsEnabled;
   private Map<String, String> tableData = EditorActionsUtil.DEFAULT_ACTIONS;
 
   public String getSystemPrompt() {
@@ -119,14 +118,6 @@ public class ConfigurationState {
     this.autoFormattingEnabled = autoFormattingEnabled;
   }
 
-  public boolean isCodeCompletionsEnabled() {
-    return codeCompletionsEnabled;
-  }
-
-  public void setCodeCompletionsEnabled(boolean codeCompletionsEnabled) {
-    this.codeCompletionsEnabled = codeCompletionsEnabled;
-  }
-
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -144,7 +135,6 @@ public class ConfigurationState {
         && methodNameGenerationEnabled == that.methodNameGenerationEnabled
         && captureCompileErrors == that.captureCompileErrors
         && autoFormattingEnabled == that.autoFormattingEnabled
-        && codeCompletionsEnabled == that.codeCompletionsEnabled
         && Objects.equals(systemPrompt, that.systemPrompt)
         && Objects.equals(commitMessagePrompt, that.commitMessagePrompt)
         && Objects.equals(tableData, that.tableData);
@@ -154,7 +144,6 @@ public class ConfigurationState {
   public int hashCode() {
     return Objects.hash(systemPrompt, commitMessagePrompt, maxTokens, temperature,
         checkForPluginUpdates, createNewChatOnEachAction, ignoreGitCommitTokenLimit,
-        methodNameGenerationEnabled, captureCompileErrors, autoFormattingEnabled,
-        codeCompletionsEnabled, tableData);
+        methodNameGenerationEnabled, captureCompileErrors, autoFormattingEnabled, tableData);
   }
 }