+++ src/main/kotlin/ee/carlrobert/codegpt/codecompletions/CodeGPTInlineCompletionProvider.kt
@@ -8,7 +8,10 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.editor.Editor
 import ee.carlrobert.codegpt.CodeGPTKeys
 import ee.carlrobert.codegpt.completions.CompletionRequestService
-import ee.carlrobert.codegpt.settings.configuration.ConfigurationSettings
+import ee.carlrobert.codegpt.settings.GeneralSettings
+import ee.carlrobert.codegpt.settings.service.ServiceType
+import ee.carlrobert.codegpt.settings.service.llama.LlamaSettings
+import ee.carlrobert.codegpt.settings.service.openai.OpenAISettings
 import ee.carlrobert.codegpt.treesitter.CodeCompletionParserFactory
 import ee.carlrobert.llm.completion.CompletionEventListener
 import kotlinx.coroutines.*
@@ -52,8 +55,13 @@ class CodeGPTInlineCompletionProvider : InlineCompletionProvider {
     }
 
     override fun isEnabled(event: InlineCompletionEvent): Boolean {
-        return event is InlineCompletionEvent.DocumentChange
-                && ConfigurationSettings.getCurrentState().isCodeCompletionsEnabled
+        val selectedService = GeneralSettings.getCurrentState().selectedService
+        val codeCompletionsEnabled = when (selectedService) {
+            ServiceType.OPENAI -> OpenAISettings.getCurrentState().isCodeCompletionsEnabled
+            ServiceType.LLAMA_CPP -> LlamaSettings.getCurrentState().isCodeCompletionsEnabled
+            else -> false
+        }
+        return event is InlineCompletionEvent.DocumentChange && codeCompletionsEnabled
     }
 
     private fun ProducerScope<InlineCompletionElement>.getCodeCompletionEventListener(